chore: Refactor to extract axe conversion to results
