+++ src/electron/platform/android/scan-results-to-unified-results.ts
@@ -3,9 +3,8 @@
 import { UnifiedResult } from 'common/types/store-data/unified-data-interface';
 import { UUIDGenerator } from 'common/uid-generator';
 import { convertAtfaScanResultsToUnifiedResults } from 'electron/platform/android/atfa-results-to-unified-results';
-import { DictionaryStringTo } from 'types/common-types';
-import { AndroidScanResults, RuleResultsData, ViewElementData } from './android-scan-results';
-import { RuleInformation } from './rule-information';
+import { convertAxeScanResultsToUnifiedResults } from 'electron/platform/android/axe-results-to-unified-results';
+import { AndroidScanResults } from './android-scan-results';
 import { RuleInformationProviderType } from './rule-information-provider-type';
 
 export type ConvertScanResultsToUnifiedResultsDelegate = (
@@ -26,81 +25,4 @@ export function convertScanResultsToUnifiedResults(
     ).concat(
         convertAtfaScanResultsToUnifiedResults(scanResults, ruleInformationProvider, uuidGenerator),
     );
-}
-
-function convertAxeScanResultsToUnifiedResults(
-    scanResults: AndroidScanResults,
-    ruleInformationProvider: RuleInformationProviderType,
-    uuidGenerator: UUIDGenerator,
-): UnifiedResult[] {
-    if (!scanResults || !scanResults.ruleResults) {
-        return [];
-    }
-
-    const viewElementLookup: DictionaryStringTo<ViewElementData> =
-        createViewElementLookup(scanResults);
-    const unifiedResults: UnifiedResult[] = [];
-
-    for (const ruleResult of scanResults.ruleResults) {
-        const ruleInformation: RuleInformation = ruleInformationProvider.getRuleInformation(
-            ruleResult.ruleId,
-        );
-
-        if (ruleInformation) {
-            unifiedResults.push(
-                createUnifiedResult(ruleInformation, ruleResult, viewElementLookup, uuidGenerator),
-            );
-        }
-    }
-
-    return unifiedResults;
-}
-
-function createViewElementLookup(
-    scanResults: AndroidScanResults,
-): DictionaryStringTo<ViewElementData> {
-    const viewElementLookup = {};
-
-    addViewElementAndChildren(viewElementLookup, scanResults.viewElementTree);
-
-    return viewElementLookup;
-}
-
-function addViewElementAndChildren(
-    viewElementLookup: DictionaryStringTo<ViewElementData>,
-    element: ViewElementData | null,
-): void {
-    if (element) {
-        viewElementLookup[element.axeViewId] = element;
-        if (element.children) {
-            for (const child of element.children) {
-                addViewElementAndChildren(viewElementLookup, child);
-            }
-        }
-    }
-}
-
-function createUnifiedResult(
-    ruleInformation: RuleInformation,
-    ruleResult: RuleResultsData,
-    viewElementLookup: DictionaryStringTo<ViewElementData>,
-    uuidGenerator: UUIDGenerator,
-): UnifiedResult {
-    const viewElement = viewElementLookup[ruleResult.axeViewId];
-    return {
-        uid: uuidGenerator(),
-        ruleId: ruleInformation.ruleId,
-        status: ruleInformation.getResultStatus(ruleResult),
-        descriptors: {
-            className: viewElement?.className,
-            boundingRectangle: viewElement?.boundsInScreen,
-            contentDescription: viewElement?.contentDescription,
-            text: viewElement?.text,
-        },
-        identifiers: {
-            identifier: viewElement?.className,
-            conciseName: viewElement?.className,
-        },
-        resolution: ruleInformation.getUnifiedResolution(ruleResult),
-    };
 }
